+++ server/server_receive/requirements.txt
@@ -1,0 +1,1 @@
+pycryptodome==3.17
