+++ server/server_receive/src/constants.py
@@ -1,0 +1,5 @@
+SEPARATOR = "<SEPARATOR>"
+BUFFER_SIZE = 4096
+CONNECT_SETTINGS_PATH = "../../settings/connectsettings.json"
+LOG_FILE_PATH = "../../data_storage/storage_info.csv"
+TIME_REGEX = r"\d{2}:\d{2}:\d{2} "
