+++ server/server_receive/src/main.py
@@ -4,12 +4,10 @@ import os
 import logupdate
 import decrypt
 
+import constants
+
 SERVER_HOST = "0.0.0.0"
 SERVER_PORT = 5001
-
-BUFFER_SIZE = 4096
-
-SEPARATOR = "<SEPARATOR>"
 
 if __name__ == "__main__":
     s = socket.socket()
@@ -17,18 +15,17 @@ if __name__ == "__main__":
     s.listen(5)
     while True:
         client_socket, address = s.accept()
-        received = client_socket.recv(BUFFER_SIZE).decode()
+        received = client_socket.recv(constants.BUFFER_SIZE).decode()
 
-        filename, filesize = received.split(SEPARATOR)
+        filename, filesize = received.split(constants.SEPARATOR)
         filename = os.path.basename(filename)
         filesize = int(filesize)
 
         logupdate.updateLogs(filename)
 
-        # TODO: add if data_storage doesn't exist create directory
         with open("../../data_storage/" + filename, "wb") as f:
             while True:
-                bytes_read = client_socket.recv(BUFFER_SIZE)
+                bytes_read = client_socket.recv(constants.BUFFER_SIZE)
                 if not bytes_read:
                     break
                 f.write(bytes_read)
@@ -37,6 +34,7 @@ if __name__ == "__main__":
             "../../data_storage/" + filename,
             "../../data_storage/" + filename.split(".")[0] + ".jpg",
         )
+        
 
     client_socket.close()
     s.close()
